아이템 정보창, 돈
- 인벤토리 슬롯에 커서를 갖다 대면 아이템 정보창이 뜸
- 몬스터를 잡으면 돈이 들어옴, 인벤토리에 현재 Coins 표기해둠.

+++ Assets/CommonRPG/Scripts/Units/MonsterDataScriptableObject.cs
@@ -44,8 +44,8 @@ namespace CommonRPG
         public float Mp;
         public float Exp;
         public float ExpTolerance;
-        public float HoldingMoney;
-        public float MoneyTolerance;
+        public int HoldingMoney;
+        public int MoneyTolerance;
         [HideInInspector]
         public List<EItemName> HoldingItems;
     }
